Update README.md
Hi, its me. The file you are going to see is my dream, which I want to come true. Yes it is a search engine, I thought the name from an anime known as 'Magical Circle Guru'. So, yeah help me to build my dream. Contribute your help in this project.